// source/src/shared/lib/enums/actions/LevelAction.ts
export enum LevelAction {
    ADD_EXAMPLE_TO_LEVEL_BY_NAME = 'ADD_EXAMPLE_TO_LEVEL_BY_NAME',
    INCRASE_SOLVED_COUNT_BY_NAME = 'INCRASE_SOLVED_COUNT_BY_NAME',
    INCRASE_UNSOLVED_COUNT_BY_NAME = 'INCRASE_UNSOLVED_COUNT_BY_NAME',
    SET_DEFAULT_STATE_IN_LOCAL_STORAGE = 'SET_DEFAULT_STATE_IN_LOCAL_STORAGE',
    CLEAR_ALL = 'CLEAR_ALL'
}